+++ rs_os/src/descriptors/mod.rs
@@ -1,4 +1,3 @@
 pub mod gdt;
 pub mod idt;
-pub mod idt2;
 pub mod segmentation;
